Add confirmation modal on local user table
Adds a delete confirmation modal when user attempting batch delete
action from local user table.

+++ src/views/AccessControl/LocalUserManagement/LocalUserManagement.vue
@@ -184,7 +184,7 @@ export default {
             {
               value: 'delete',
               enabled: user.UserName === 'root' ? false : true,
-              title: this.$t('pageLocalUserManagement.deleteUser')
+              title: this.$tc('pageLocalUserManagement.deleteUser')
             }
           ],
           ...user
@@ -215,8 +215,8 @@ export default {
             user: user.username
           }),
           {
-            title: this.$t('pageLocalUserManagement.deleteUser'),
-            okTitle: this.$t('pageLocalUserManagement.deleteUser')
+            title: this.$tc('pageLocalUserManagement.deleteUser'),
+            okTitle: this.$tc('pageLocalUserManagement.deleteUser')
           }
         )
         .then(deleteConfirmed => {
@@ -250,13 +250,34 @@ export default {
     onBatchAction(action) {
       switch (action) {
         case 'delete':
-          this.$store
-            .dispatch('localUsers/deleteUsers', this.selectedRows)
-            .then(messages => {
-              messages.forEach(({ type, message }) => {
-                if (type === 'success') this.successToast(message);
-                if (type === 'error') this.errorToast(message);
-              });
+          this.$bvModal
+            .msgBoxConfirm(
+              this.$tc(
+                'pageLocalUserManagement.modal.batchDeleteConfirmMessage',
+                this.selectedRows.length
+              ),
+              {
+                title: this.$tc(
+                  'pageLocalUserManagement.deleteUser',
+                  this.selectedRows.length
+                ),
+                okTitle: this.$tc(
+                  'pageLocalUserManagement.deleteUser',
+                  this.selectedRows.length
+                )
+              }
+            )
+            .then(deleteConfirmed => {
+              if (deleteConfirmed) {
+                this.$store
+                  .dispatch('localUsers/deleteUsers', this.selectedRows)
+                  .then(messages => {
+                    messages.forEach(({ type, message }) => {
+                      if (type === 'success') this.successToast(message);
+                      if (type === 'error') this.errorToast(message);
+                    });
+                  });
+              }
             });
           break;
         case 'enable':
